+++ src/lib/algorithms/reachability/helperMethods.h
@@ -278,7 +278,7 @@ namespace hypro{
             
             for(int i=0; i<m.rows(); i++)
             {
-                result[i] = m(i,0).toDouble();
+                result[i] = m(i,0);
             }
             
             return result;
@@ -293,7 +293,7 @@ namespace hypro{
             
             for(int i=0; i<v.size(); i++)
             {
-                result[i] = v(i).toDouble();
+                result[i] = v(i);
             }
             
             return result;
